chore: move OverlayModule and related providers into index file
Moves the `OverlayModule` and the `OVERLAY_PROVIDERS` into the `index.ts`, as opposed to the `overlay-directives.ts`, in order to match the setup in the rest of the modules.
