+++ src/lib/core/overlay/overlay-directives.ts
@@ -7,7 +7,6 @@
  */
 
 import {
-    NgModule,
     Directive,
     EventEmitter,
     TemplateRef,
@@ -21,7 +20,7 @@ import {
     OnChanges,
     SimpleChanges,
 } from '@angular/core';
-import {Overlay, OVERLAY_PROVIDERS} from './overlay';
+import {Overlay} from './overlay';
 import {OverlayRef} from './overlay-ref';
 import {TemplatePortal} from '../portal/portal';
 import {OverlayState} from './overlay-state';
@@ -29,7 +28,6 @@ import {
     ConnectionPositionPair,
     ConnectedOverlayPositionChange
 } from './position/connected-position';
-import {PortalModule} from '../portal/portal-directives';
 import {ConnectedPositionStrategy} from './position/connected-position-strategy';
 import {Directionality, Direction} from '../bidi/index';
 import {Scrollable} from './scroll/scrollable';
@@ -37,7 +35,6 @@ import {ScrollStrategy} from './scroll/scroll-strategy';
 import {coerceBooleanProperty} from '../coercion/boolean-property';
 import {ESCAPE} from '../keyboard/keycodes';
 import {Subscription} from 'rxjs/Subscription';
-import {ScrollDispatchModule} from './scroll/index';
 
 
 /** Default set of positions for the overlay. Follows the behavior of a dropdown. */
@@ -326,12 +323,3 @@ export class ConnectedOverlayDirective implements OnDestroy, OnChanges {
     });
   }
 }
-
-
-@NgModule({
-  imports: [PortalModule, ScrollDispatchModule],
-  exports: [ConnectedOverlayDirective, OverlayOrigin, ScrollDispatchModule],
-  declarations: [ConnectedOverlayDirective, OverlayOrigin],
-  providers: [OVERLAY_PROVIDERS],
-})
-export class OverlayModule {}
