+++ src/lib/core/overlay/overlay.ts
@@ -12,14 +12,12 @@ import {
   ApplicationRef,
   Injector,
   NgZone,
-  Provider,
 } from '@angular/core';
 import {OverlayState} from './overlay-state';
 import {DomPortalHost} from '../portal/dom-portal-host';
 import {OverlayRef} from './overlay-ref';
 import {OverlayPositionBuilder} from './position/overlay-position-builder';
-import {VIEWPORT_RULER_PROVIDER} from './position/viewport-ruler';
-import {OverlayContainer, OVERLAY_CONTAINER_PROVIDER} from './overlay-container';
+import {OverlayContainer} from './overlay-container';
 import {ScrollStrategy, ScrollStrategyOptions} from './scroll/index';
 
 
@@ -99,11 +97,3 @@ export class Overlay {
     return new OverlayRef(portalHost, pane, state, scrollStrategy, this._ngZone);
   }
 }
-
-/** Providers for Overlay and its related injectables. */
-export const OVERLAY_PROVIDERS: Provider[] = [
-  Overlay,
-  OverlayPositionBuilder,
-  VIEWPORT_RULER_PROVIDER,
-  OVERLAY_CONTAINER_PROVIDER,
-];
