+++ src/lib/sidenav/index.ts
@@ -10,7 +10,7 @@ import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {MdCommonModule} from '../core';
 import {A11yModule} from '../core/a11y/index';
-import {OverlayModule} from '../core/overlay/overlay-directives';
+import {OverlayModule} from '../core/overlay/index';
 import {MdSidenav, MdSidenavContainer} from './sidenav';
